refactor: postcard와 swiper 컴포넌트 파일 분리 #7
