+++ .gitignore
@@ -26,6 +26,10 @@ yarn-error.log*
 
 # local env files
 .env*.local
+.env
+.env.local
+.env.production
+.env.development
 
 # vercel
 .vercel
